protein_proj/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.3.0
xmltodict==0.14.2
